+++ app/src/main/java/com/xiaojinzi/componentdemo/App.java
@@ -44,7 +44,6 @@ public class App extends Application {
         if (BuildConfig.DEBUG) {
             ModuleManager.getInstance().check();
         }
-
     }
 
 }